requirements.txt
chromadb
fastapi
PyPDF2
joblib
PyYAML
langchain-openai
openai
scikit-learn
transformers
requests
httpx
uvicorn
python-multipart
unstructured
sentence-transformers
faiss-cpu
pandas
pdfminer
pdfminer.six  
fastapi  
azure-cognitiveservices-language-luis  
langchain  
langchain_community  
pathlib  
certifi  
openai  
nltk  
python-dotenv  